+++ CHANGELOG.md
@@ -24,6 +24,7 @@
     - also add ``DIRSTACK`` global variable
 - add ``Array#removeRange`` method
     - remove elements by specified range
+- add ``String#quote`` method for generate string that is used as command argument
 
 #### Module